Bring back YamlParserTest
